GH-1755: move parser / writer settings to specific packages Signed-off-by:Bart Hanssens <bart.hanssens@bosa.fgov.be>
