+++ core/rio/jsonld/src/main/java/org/eclipse/rdf4j/rio/jsonld/JSONLDSettings.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
+ * Copyright (c) 2022 Eclipse RDF4J contributors.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Distribution License v1.0
@@ -8,9 +8,12 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *******************************************************************************/
-package org.eclipse.rdf4j.rio.helpers;
+package org.eclipse.rdf4j.rio.jsonld;
 
 import org.eclipse.rdf4j.rio.RioSetting;
+import org.eclipse.rdf4j.rio.helpers.BooleanRioSetting;
+import org.eclipse.rdf4j.rio.helpers.ClassRioSetting;
+import org.eclipse.rdf4j.rio.helpers.RioSettingImpl;
 
 import com.github.jsonldjava.core.DocumentLoader;
 
@@ -19,6 +22,8 @@ import com.github.jsonldjava.core.DocumentLoader;
  *
  * @author Peter Ansell
  * @see <a href="http://json-ld.org/spec/latest/json-ld-api/#data-structures">JSONLD Data Structures</a>
+ *
+ * @since 4.3.0
  */
 public class JSONLDSettings {
 
@@ -31,20 +36,14 @@ public class JSONLDSettings {
 	 * Can be overridden by setting system property {@code org.eclipse.rdf4j.rio.jsonld.compact_arrays}.
 	 *
 	 * @see <a href="http://json-ld.org/spec/latest/json-ld-api/#data-structures">JSONLD Data Structures</a>
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#COMPACT_ARRAYS} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<Boolean> COMPACT_ARRAYS = new BooleanRioSetting(
 			"org.eclipse.rdf4j.rio.jsonld.compact_arrays", "Compact arrays", Boolean.TRUE);
 
 	/**
 	 * If specified, it is used to retrieve remote documents and contexts; otherwise the processor's built-in loader is
 	 * used.
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#DOCUMENT_LOADER} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<DocumentLoader> DOCUMENT_LOADER = new ClassRioSetting<>(
 			"org.eclipse.rdf4j.rio.jsonld.document_loader", "Document loader", null);
 
@@ -58,10 +57,7 @@ public class JSONLDSettings {
 	 * Can be overridden by setting system property {@code org.eclipse.rdf4j.rio.jsonld.optimize}.
 	 *
 	 * @see <a href="http://json-ld.org/spec/latest/json-ld-api/#data-structures">JSONLD Data Structures</a>
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#OPTIMIZE} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<Boolean> OPTIMIZE = new BooleanRioSetting("org.eclipse.rdf4j.rio.jsonld.optimize",
 			"Optimize output", Boolean.FALSE);
 
@@ -76,10 +72,7 @@ public class JSONLDSettings {
 	 * Can be overridden by setting system property {@code org.eclipse.rdf4j.rio.jsonld.produce_generalized_rdf}.
 	 *
 	 * @see <a href="http://json-ld.org/spec/latest/json-ld-api/#data-structures">JSONLD Data Structures</a>
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#PRODUCE_GENERALIZED_RDF} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<Boolean> PRODUCE_GENERALIZED_RDF = new BooleanRioSetting(
 			"org.eclipse.rdf4j.rio.jsonld.produce_generalized_rdf", "Produce generalized RDF", Boolean.FALSE);
 
@@ -93,10 +86,7 @@ public class JSONLDSettings {
 	 * Can be overridden by setting system property {@code org.eclipse.rdf4j.rio.jsonld.use_native_types}.
 	 *
 	 * @see <a href="http://json-ld.org/spec/latest/json-ld-api/#data-structures">JSONLD Data Structures</a>
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#USE_NATIVE_TYPES} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<Boolean> USE_NATIVE_TYPES = new BooleanRioSetting(
 			"org.eclipse.rdf4j.rio.jsonld.use_native_types", "Use Native JSON Types", Boolean.FALSE);
 
@@ -109,10 +99,7 @@ public class JSONLDSettings {
 	 * Can be overridden by setting system property {@code org.eclipse.rdf4j.rio.jsonld.use_rdf_type}.
 	 *
 	 * @see <a href="http://json-ld.org/spec/latest/json-ld-api/#data-structures">JSONLD Data Structures</a>
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#USE_RDF_TYPE} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<Boolean> USE_RDF_TYPE = new BooleanRioSetting(
 			"org.eclipse.rdf4j.rio.jsonld.use_rdf_type", "Use RDF Type", Boolean.FALSE);
 
@@ -122,10 +109,7 @@ public class JSONLDSettings {
 	 * Defaults to {@link JSONLDMode#EXPAND} to provide maximum RDF compatibility.
 	 *
 	 * @see <a href="http://json-ld.org/spec/latest/json-ld-api/#features">JSONLD Features</a>
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#JSONLD_MODE} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<JSONLDMode> JSONLD_MODE = new RioSettingImpl<>(
 			"org.eclipse.rdf4j.rio.jsonld_mode", "JSONLD Mode", JSONLDMode.EXPAND);
 
@@ -135,10 +119,7 @@ public class JSONLDSettings {
 	 * Default to false
 	 * <p>
 	 * Can be overridden by setting system property {@code org.eclipse.rdf4j.rio.jsonld.hierarchical_view}.
-	 *
-	 * @deprecated Use {@link org.eclipse.rdf4j.rio.jsonld.JSONLDSettings#HIERARCHICAL_VIEW} instead.
 	 */
-	@Deprecated(since = "4.3.0", forRemoval = true)
 	public static final RioSetting<Boolean> HIERARCHICAL_VIEW = new BooleanRioSetting(
 			"org.eclipse.rdf4j.rio.jsonld.hierarchical_view", "Hierarchical representation of the JSON", Boolean.FALSE);
